+++ config/loud-jobs.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for Datashaman\LoudJobs/LoudJobs
+return [
+
+];