+++ src/Commands/LoudJobsCommand.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace VendorName\Skeleton\Commands;
+namespace Datashaman\LoudJobs\Commands;
 
 use Illuminate\Console\Command;
 
-class SkeletonCommand extends Command
+class LoudJobsCommand extends Command
 {
-    public $signature = 'skeleton';
+    public $signature = 'loud-jobs';
 
     public $description = 'My command';