+++ src/LoudJobs.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Datashaman\LoudJobs;
+
+class LoudJobs {}