+++ src/LoudJobsServiceProvider.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace VendorName\Skeleton;
+namespace Datashaman\LoudJobs;
 
+use Datashaman\LoudJobs\Commands\LoudJobsCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
-use VendorName\Skeleton\Commands\SkeletonCommand;
 
-class SkeletonServiceProvider extends PackageServiceProvider
+class LoudJobsServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
@@ -16,10 +16,10 @@ class SkeletonServiceProvider extends PackageServiceProvider
          * More info: https://github.com/spatie/laravel-package-tools
          */
         $package
-            ->name('skeleton')
+            ->name('loud-jobs')
             ->hasConfigFile()
             ->hasViews()
-            ->hasMigration('create_migration_table_name_table')
-            ->hasCommand(SkeletonCommand::class);
+            ->hasMigration('create_loud_jobs_table')
+            ->hasCommand(LoudJobsCommand::class);
     }
 }